+++ gdsfactory/simulation/gtidy3d/get_simulation.py
@@ -1,10 +1,11 @@
 """Returns tidy3d simulation from gdsfactory Component."""
 import warnings
-from typing import Dict, Optional, Union
+from typing import Dict, Optional
 
 import matplotlib as mpl
 import matplotlib.pyplot as plt
 import numpy as np
+import pydantic
 import tidy3d as td
 
 import gdsfactory as gf
@@ -16,16 +17,22 @@ from gdsfactory.simulation.gtidy3d.materials import get_index, get_medium
 from gdsfactory.tech import LAYER_STACK, LayerStack
 from gdsfactory.types import Float2
 
-MATERIAL_NAME_TO_TIDY3D = {
+# not dispersive materials have a constant index
+MATERIAL_NAME_TO_TIDY3D_INDEX = {
     "si": 3.47,
     "sio2": 1.44,
     "sin": 2.0,
-    # "si": "cSi",
-    # "sio2": "SiO2",
-    # "sin": "Si3N4",
+}
+
+# dispersive materials
+MATERIAL_NAME_TO_TIDY3D_NAME = {
+    "si": "cSi",
+    "sio2": "SiO2",
+    "sin": "Si3N4",
 }
 
 
+@pydantic.validate_arguments
 def get_simulation(
     component: Component,
     port_extension: Optional[float] = 4.0,
@@ -50,7 +57,9 @@ def get_simulation(
     plot_modes: bool = False,
     num_modes: int = 2,
     run_time_ps: float = 10.0,
-    material_name_to_tidy3d: Dict[str, Union[float, str]] = MATERIAL_NAME_TO_TIDY3D,
+    dispersive: bool = False,
+    material_name_to_tidy3d_index: Dict[str, float] = MATERIAL_NAME_TO_TIDY3D_INDEX,
+    material_name_to_tidy3d_name: Dict[str, str] = MATERIAL_NAME_TO_TIDY3D_NAME,
     is_3d: bool = True,
 ) -> td.Simulation:
     r"""Returns Simulation object from gdsfactory.component
@@ -118,8 +127,14 @@ def get_simulation(
         plot_modes: plot source modes.
         num_modes: number of modes to plot.
         run_time_ps: make sure it's sufficient for the fields to decay.
-            defaults to 10ps and counts on the automatic shutoff to stop earlier if needed.
-
+            defaults to 10ps and counts on automatic shutoff to stop earlier if needed.
+        dispersive: False uses constant refractive index materials.
+            True adds wavelength depending materials.
+            Dispersive materials require more computation.
+        material_name_to_tidy3d_index: not dispersive materials have a constant index.
+        material_name_to_tidy3d_name: dispersive materials have a wavelength
+            dependent index. Maps layer_stack names with tidy3d material database names.
+        is_3d: if False, does not consider Z dimension for faster simulations.
 
     .. code::
 
@@ -136,6 +151,11 @@ def get_simulation(
     layer_to_material = layer_stack.get_layer_to_material()
     layer_to_zmin = layer_stack.get_layer_to_zmin()
     # layer_to_sidewall_angle = layer_stack.get_layer_to_sidewall_angle()
+
+    if dispersive:
+        material_name_to_tidy3d = material_name_to_tidy3d_name
+    else:
+        material_name_to_tidy3d = material_name_to_tidy3d_index
 
     assert isinstance(
         component, Component